requirements.txt
tkinterdnd2==0.3.0
pillow>=9.0.0